Fixing build errors
