+++ locales
@@ -1,0 +1,1 @@
+public/locales
