+++ src/app/contact/page.tsx
@@ -1,6 +1,6 @@
-import ContactPage from "../components/contact/Entry"
-import Footer from "../components/landing/Footer"
-import Header from "../components/landing/Header"
+import ContactPage from "../../components/contact/Entry"
+import Footer from "../../components/landing/Footer"
+import Header from "../../components/landing/Header"
 
 const Home = () => {
   return (
