+++ src/app/new/[type]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import React from 'react'
-import PaoCreationPage from '@/app/components/new/PAOCreation'
+import PaoCreationPage from '@/components/new/PAOCreation'
 
 const Home = function() {
   return <div><PaoCreationPage /></div>
