+++ src/app/pao/[id]/layout.tsx
@@ -1,6 +1,6 @@
-import DashboardHeader from "@/app/components/getstarted/DashboardHeader";
-import Footer from "@/app/components/landing/Footer";
-import AppWalletProvider from "@/app/components/AppWalletProvider";
+import DashboardHeader from "@/components/getstarted/DashboardHeader";
+import Footer from "@/components/landing/Footer";
+import AppWalletProvider from "@/components/AppWalletProvider";
 
 export default function NewLayout({ children }: { children: React.ReactNode }) {
   return (
