+++ src/components/about/Entry.tsx
