+++ src/components/about/Hero.tsx
