+++ src/components/landing/Entry.tsx
