+++ src/components/landing/Hero.tsx
